refactor(mobile): extract side effects from screens into feature hooks (OJD-1522)
Component layer cleanup: components and app routes no longer reach into
react-query or db services directly. All onSuccess / invalidate /
resetQueries / prefetch effects now live in hooks.

Specifics:

- features/auth/hooks/use-login.ts: prefetchOfflineData now runs inside
  each useMutation's onSuccess (gated on result.error for OTP verify).
  app/(auth)/login.tsx loses useQueryClient + 3 inline prefetch calls.
- features/auth/hooks/use-logout.ts (new): wraps useSession.signOut with
  queryClient.resetQueries so callers don't have to coordinate them.
  app/(tabs)/profile.tsx now just calls useLogout() and drops
  useQueryClient.
- features/recent-measurements/hooks/use-prune-expired-measurements.ts
  (new): boot-time effect that prunes expired SQLite rows and
  invalidates the measurements query. app/(tabs)/_layout.tsx no longer
  imports useQueryClient or pruneExpiredMeasurements directly.

Audited but unchanged (no actual side-effect-in-render violations):

- features/measurement-flow/.../question-node/question-node.tsx: 220 LOC
  is JSX styling; toast calls are inside event handlers (allowed). Size
  reduction is a component-extraction concern, not layer-boundary work.
- features/recent-measurements/components/measurement-questions-modal.tsx:
  283 LOC is similarly all JSX. updateMeasurementComment is consumed
  via useMeasurements (a hook), not invoked through useQueryClient.

Verification: typecheck 0 errors, lint clean, 362/362 tests pass. grep
for useQueryClient/invalidateQueries inside features/*/components,
features/*/screens, and src/app returns 0 hits.
