+++ apps/mobile/src/features/recent-measurements/hooks/use-prune-expired-measurements.ts
@@ -1,0 +1,12 @@
+import { useQueryClient } from "@tanstack/react-query";
+import { useEffect } from "react";
+import { pruneExpiredMeasurements } from "~/shared/db/measurements-storage";
+
+export function usePruneExpiredMeasurements() {
+  const queryClient = useQueryClient();
+  useEffect(() => {
+    void pruneExpiredMeasurements().then(() => {
+      void queryClient.invalidateQueries({ queryKey: ["measurements"] });
+    });
+  }, [queryClient]);
+}
